Fix: カードのユーザー名取得処理でユーザー名が更新されないバグを直した
Fix:
前回第二引数にphotoSnapListを格納したuseEffectで、更新のたびにmapで回してユーザー名を取得するやり方を取ったが、この方法だとカードに表示されるユーザー名とカードを飛んだ先のユーザー名が異なる時があると言う問題が発覚した。
- return文のphotoSnapList.map内でgetUserNameを使用していた時は、無限ループの問題はあったものの適切にユーザー名を取得できていた
- カードを飛んだ先では適切なユーザー名が表示されることから、getUserName(data.uid)とすれば確実に求めるユーザー名が取得できる

以上より、内部を別コンポーネントに移せば、適切にユーザー名を取得ができ、かつコンポーネントが別なので無限レンダリングバグも起きないと考えた。
結果、恐らくバグは解消されたものと思われる。
